[libc++][NFC] Merge add_{r,l}value_reference.h
