+++ libcxx/include/CMakeLists.txt
@@ -778,9 +778,8 @@ set(files
   __tuple/tuple_size.h
   __tuple/tuple_types.h
   __type_traits/add_cv_quals.h
-  __type_traits/add_lvalue_reference.h
   __type_traits/add_pointer.h
-  __type_traits/add_rvalue_reference.h
+  __type_traits/add_reference.h
   __type_traits/aligned_storage.h
   __type_traits/aligned_union.h
   __type_traits/alignment_of.h
