+++ libcxx/include/experimental/memory
@@ -57,8 +57,8 @@ public:
 #  include <__cstddef/size_t.h>
 #  include <__functional/hash.h>
 #  include <__functional/operations.h>
-#  include <__type_traits/add_lvalue_reference.h>
 #  include <__type_traits/add_pointer.h>
+#  include <__type_traits/add_reference.h>
 #  include <__type_traits/common_type.h>
 #  include <__type_traits/enable_if.h>
 #  include <__type_traits/is_convertible.h>
